Rename ObjectCopyUtil to AblyObjectsCopyUtil
