+++ core/src/main/java/io/ably/lib/rest/AblyBase.java
@@ -31,10 +31,10 @@ import io.ably.lib.types.Param;
 import io.ably.lib.types.PublishResponse;
 import io.ably.lib.types.Stats;
 import io.ably.lib.types.StatsReader;
+import io.ably.lib.util.AblyObjectsCopyUtil;
 import io.ably.lib.util.Crypto;
 import io.ably.lib.util.InternalMap;
 import io.ably.lib.util.Log;
-import io.ably.lib.util.ObjectCopyUtil;
 import io.ably.lib.util.PlatformAgentProvider;
 import io.ably.lib.util.Serialisation;
 
@@ -89,7 +89,7 @@ public abstract class AblyBase<
         if (options instanceof DebugOptions) {
             this.options = options;
         } else {
-            this.options = ObjectCopyUtil.copy(options);
+            this.options = AblyObjectsCopyUtil.copy(options);
         }
 
         /* process options */
