+++ core/src/main/java/io/ably/lib/util/AblyObjectsCopyUtil.java
